passed the 404 delete test on the server side.

+++ server/src/items/items.controller.ts
@@ -1,4 +1,6 @@
 import { Body, ConsoleLogger, Controller, Delete, Get, HttpCode, HttpException, HttpStatus, Param, ParseFloatPipe, ParseIntPipe, Post, Query, Req, Res, UsePipes, ValidationPipe } from '@nestjs/common';
+import { isArray } from 'class-validator';
+import { isEmpty } from 'cypress/types/lodash';
 import { ItemClassDto } from './ClassDTO/itemsClass.dto';
 import { Items_Interface } from './items-Model/items-model';
 import { ItemsService } from './items.service';
@@ -31,13 +33,19 @@ export class ItemsController {
        //dto --> according to the doc. defines how the data is sent over the network, we want to ensure that the received data matches this DTO
     @Post('/item')
     @UsePipes(ValidationPipe)
-    createItem(@Body() Create_Item: ItemClassDto ){
+    createItem(@Body() Create_Item: ItemClassDto ):Items_Interface{
        return this.ItemsService.CreateItem(Create_Item);
     }
 
     @Delete('/item/:id')
     removeItemById(@Param('id',ParseIntPipe) id: number):void{
-      this.ItemsService.removeItemById(id);
+      const found = this.ItemsService.findItemsById(id) 
+      if(isNaN(found.id)){
+         throw new HttpException('item not found!', HttpStatus.NOT_FOUND)
+      }
+      else {
+         this.ItemsService.removeItemById(id);
       throw new HttpException('Item Deleted!', HttpStatus.NO_CONTENT);
+   }
     }
 }
